+++ pingresp_test.go
@@ -1,0 +1,9 @@
+package mqtt
+
+import "testing"
+
+func TestPingResp(t *testing.T) {
+	p := NewPingResp()
+
+	testControlPacket(t, &p)
+}